+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Change Log
 
-## v1.0.19 ~ 11-5-2020
+## v1.0.19 ~ 11-7-2020
 
   * New Module: Trigs
   * Tree module now resizable (Thanks to SteveRussell33)
   * ThingThing module now resizable (Thanks to me)
+  * Bouncy Balls multiplier lower limit now less than 1 so you can slow down the ball and move it with cv.
 
 ## v1.0.18 ~ 11-2-2020